# apps/backend/api/swag-docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response:
    properties:
      email:
        type: string
      familyName:
        type: string
      givenName:
        type: string
      id:
        type: string
      name:
        type: string
      orcid:
        type: string
      person_id:
        type: string
    type: object
  github_com_SURF-Innovatie_MORIS_internal_auth.BackendError:
    properties:
      code:
        example: 400
        type: integer
      errors:
        description: Can be map[string]string or []string or null
      message:
        description: Custom message
        example: Detailed error description
        type: string
      status:
        example: Bad Request
        type: string
    type: object
...
  internal_handler_custom.LinkORCIDRequest:
    properties:
      code:
        example: authentication_code_from_orcid
        type: string
    type: object
  internal_handler_custom.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secretpassword
        type: string
    type: object
  internal_handler_custom.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
    type: object
  internal_handler_custom.ORCIDAuthURLResponse:
    properties:
      url:
        example: https://orcid.org/oauth/authorize?...
        type: string
    type: object
  internal_handler_custom.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: secretpassword
        type: string
    type: object
  internal_handler_custom.StatusResponse:
    properties:
      status:
        example: ok
        type: string
      timestamp:
        example: "2025-11-12T10:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MORIS
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MORIS
  version: "1.0"
paths:
  /admin/users/list:
    get:
      consumes:
      - application/json
      description: Returns a list of all users - requires admin role
      produces:
      - application/json
      responses:
        "200":
          description: Admin user list
          schema:
            type: string
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - admin
  /auth/orcid/link:
    post:
      consumes:
      - application/json
      description: Links an ORCID ID to the authenticated user's account
      parameters:
      - description: ORCID link request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_custom.LinkORCIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_custom.StatusResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "409":
          description: ORCID ID already linked
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      security:
      - BearerAuth: []
      summary: Link ORCID ID
      tags:
      - auth
  /auth/orcid/unlink:
    post:
      consumes:
      - application/json
      description: Unlinks the ORCID ID from the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_custom.StatusResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      security:
      - BearerAuth: []
      summary: Unlink ORCID ID
      tags:
      - auth
  /auth/orcid/url:
    get:
      consumes:
      - application/json
      description: Returns the URL to redirect the user to for ORCID authentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_custom.ORCIDAuthURLResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      security:
      - BearerAuth: []
      summary: Get ORCID authorization URL
      tags:
      - auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_handler_custom.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_custom.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      summary: Login user
      tags:
      - auth
...
  /profile:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided credentials
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_handler_custom.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_api_userdto.Response'
        "400":
          description: Invalid request body or missing fields
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_SURF-Innovatie_MORIS_internal_auth.BackendError'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" then a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"